Shared selectbox (#2795)
