+++ frontend/src/components/shared/Dropdown/index.tsx
@@ -15,4 +15,5 @@
  * limitations under the License.
  */
 
-export { default } from "./VirtualDropdown"
+export { default as VirtualDropdown } from "./VirtualDropdown"
+export { default } from "./Selectbox"
